requirements.txt
pytest-lazy-fixture==0.5.2
requests==2.21.0
websockets==7.0
xmltodict==0.12.0
selene==1.0.0a13
flaky==3.6.1
pre-commit==2.6.0
flake8==3.8.3
allure-pytest==2.10.0
selenium==4.4.3
pytest==7.1.3
python-dotenv==0.20.0
environs==9.5.0
webdriver-manager==3.8.4